Helper Functions

These are subroutines that need to be used multiple times, but don't seem important enough to be their own function at the moment. Some of them may one day deserve to be promoted to proper functions in the future.

OperatorMonotoneCorrelationTools._makereal — Function

_makereal(x :: Vector, tol=1e-10)

Eigenvalues need to be real, but numerical methods sometimes have some small error. This checks if there will be an issue and if there won't be, it returns the vector with real elements.

OperatorMonotoneCorrelationTools._depolkraus — Function

_depolkraus(q)

This function returns the Kraus operators of the qubit depolarizing channel with parameter $q.$

OperatorMonotoneCorrelationTools._transposekraus — Function

_transposekraus(d)

This function returns the Kraus operators for the transpose map acting on dimension $d.$

OperatorMonotoneCorrelationTools._parallelchan — Function

_parallelchan(Ak,Bk)

This function returns the Kraus operators of a channel being ran twice in parallel.