$J_{f,\\sigma }^{p}$ Functions

These functions calculate quantities directly relevant to or induced by the $J_{f,\\sigma}^{p}$ Operator.

Direct Functions of $J_{f,\\sigma }^{p}$

OperatorMonotoneCorrelationTools.Jfpsigma — Function

Jfpsigma(Y,sigma,p,f,f0,fpinf)

This function computes $\mathbf{J}_{f,\sigma}^{p}(Y)$. It expects the input in the computational basis and returns the matrix in the computational basis.

OperatorMonotoneCorrelationTools.Jfpsigmachoi — Function

Jfpsigmachoi(σ,p,f,f0,fpinf)

This function returns the Choi operator of $\mathbf{J}_{f,\sigma}^{p}.$ We note this has a specific function for obtaining the Choi operator to force the user to consider $p,f,f(0+),f'(+\inf).$

OperatorMonotoneCorrelationTools.innerproductf — Function

innerproductf(X,Y,sigma,p,f,f0,fpinf)

This function computes $\langle X , Y \rangle_{\mathbf{J}^{p}_{f,\sigma}}$. Note that it does not check the function $f$ is an operator monotone function.

Contraction Coefficient Functions

OperatorMonotoneCorrelationTools.SchReversalMap — Function

SchReversalMap(X,Ak,Bk,σ,f,f0,fpinf)

Applies the Schrodinger reversal map to X according to $f$,$\mathcal{E}$, and $σ.$ $\mathcal{E}$ is presumed to be provided in its Kraus operator form. It is assumed all inputs are expressed in the computational basis.

OperatorMonotoneCorrelationTools.getcontractioncoeff — Function

getcontractioncoeff(Ak, Bk, σ, f, f0, fpinf)

This returns the contraction coefficient $\eta_{\chi^{2}_{f}}(\mathcal{E},\sigma)$ for a full rank input state $\sigma$ and symmetric-inducing operator monotone function $f.$

Maximal Correlation Coefficients

OperatorMonotoneCorrelationTools.qmaxcorrcoeff — Function

qmaxcorrcoeff(ρA::Matrix, Ak::Vector, Bk::Vector, f, f0, fpinf)

This function computes the maximal correlation coefficient $\mu_{f}(\rho_{AB})$ when given $\rho_{A}$ and the Kraus operators of $\mathcal{E}$ such that $\rho_{AB} = (\text{id}_{A} \otimes \mathcal{E})(\psi_{\rho_{A}})$ where $\psi_{\rho_{A}}$ is the canonical purification of $\rho_{A}$.

OperatorMonotoneCorrelationTools.qmaxlincorrcoeff — Function

qmaxlincorrcoeff(ρA::Matrix, Ak::Vector, Bk::Vector,k)

This function computes the maximal correlation coefficient $\mu_{f_{k}}(\rho_{AB})$ for $f_{k}(x) = x^{k}.$ Currently it requires that it is given $\rho_{A}$ and the Kraus operators of $\mathcal{E}$ such that $\rho_{AB} = (\text{id}_{A} \otimes \mathcal{E})(\psi_{\rho_{A}})$ where $\psi_{\rho_{A}}$ is the canonical purification of $\rho_{A}$.

Ancillary Functions

OperatorMonotoneCorrelationTools.perspective — Function

perspective(x,y,f,f0,fpinf)

For a given function f, this computes the perspective function

\[ P_{f}(x,y) \coloneqq \begin{cases} yf(x/y) & x,y > 0 , \\ yf(0^{+}) & x = 0 , \\ xf'(+\infty) & y = 0 \end{cases}\]

where $0f(0/0) \coloneqq 0$, $0\cdot \infty \coloneqq 0$,

\[ f(0^{+}) \coloneqq \lim_{x \downarrow 0} f(x) , \quad \text{and} \quad f'(+\infty) \coloneqq \lim_{x \to +\infty} \frac{f(x)}{x} . \]

We note that we allow one to control f0 and fpinf. The function will not work if these values are wrong. We assume you will put Inf (resp. -Inf) if f0 or fpinf is infinite.

OperatorMonotoneCorrelationTools.getONB — Function

getONB(σ,p,f,f0,fpinf)

This function performs the (modified) Gram Schmidt process for the inner product spaces $\langle X,Y \rangle_{\mathbf{J}_{f,\sigma}^{p}}$ considered in the paper.

Note inputs need to be in computational basis and are returned in computational basis as the inner product value is a number and thus does not change the basis here.