Random Objects

These functions generate random instances of certain quantum objects.

OperatorMonotoneCorrelationTools.Haarrandomunitary — Function

RandomUnitary(d)

This function returns a unitary of dimension d according to the Haar measure. The construction follows "How to generate a random unitary matrix" by Maris Ozols.

OperatorMonotoneCorrelationTools.hsrandomstate — Function

hsrandomstate(d::Int,k::Int=d, re=false)

Draws a density matrix according to the $\mu_{n,k}$ distribution. The method of construction follows Lemma 1 of "Asymptotics of random density matrices" by Ion Nechita. In the case the variable re is set to true, the function generates a random state with only real entries. No analytic structure about this measure is guaranteed.

OperatorMonotoneCorrelationTools.randomquantumchannel — Function

randomquantumchannel(dA,dB)

Returns the Choi state of a randomquantum channel by drawing $\rho_{AB}$ according to the Hilbert-Schmidt measure and then returning $\rho_{A}^{-1/2}\rho_{AB}\rho_{A}^{-1/2}$. If re is set to true, it generates the quantum channel from a quantum state with real values.